% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_support.R
\name{prepare_support}
\alias{prepare_support}
\title{create file to send to eoda R Support}
\usage{
prepare_support(rfile)
}
\arguments{
\item{rfile}{character string standalone R script which produces Error}
}
\value{
creates html file with same name of rfile in working directory
}
\description{
create file to send to eoda R Support
}
\examples{
prepare_support("test_file.R")

}

