% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_object.R
\name{recreate_object}
\alias{recreate_object}
\title{create code for reproduced data}
\usage{
recreate_object(x)
}
\value{
string with code to recreate data
}
\description{
create code for reproduced data
}
\examples{
recreate_object("mtcars")
}

