% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_data.R
\name{reproduce_data}
\alias{reproduce_data}
\title{create file with code for reproduced data to send to eoda R Support}
\usage{
reproduce_data()
}
\value{
creates r file "reproduced_data.R" in working directory
}
\description{
create file with code for reproduced data to send to eoda R Support
}
\examples{
reproduce_data()

}

